requirements.txt
torch
transformers
lmdb
gradio
pandas
tqdm
numpy
faiss-cpu
flask